+++ lib/src/private/sub_extension/context/context_size_extension.dart
@@ -1,9 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:kartal/src/sub_extension/context/index.dart';
+import 'package:kartal/src/private/sub_extension/context/index.dart';
 import 'package:kartal/src/widget/sized_box/space_sized_height_box.dart';
 import 'package:kartal/src/widget/sized_box/space_sized_width_box.dart';
 
-class _ContextSizeExtension {
+extension SizedBoxExtension on BuildContext {
+  _ContextSizeExtension get sized => _ContextSizeExtension(this);
+}
+
+final class _ContextSizeExtension {
   _ContextSizeExtension(BuildContext context) : _context = context;
   final BuildContext _context;
 
@@ -55,51 +59,5 @@ class _ContextSizeExtension {
   double dynamicWidth(double val) => width * val;
 
   /// Calculates and returns a dynamic height value based on the provided [val] and the current widget's height.
-  double dynamicHeight(double val) => height * val;
-}
-
-// Deprecated extension methods for [BuildContext] to create [Radius] objects.
-extension SizedBoxExtension on BuildContext {
-  _ContextSizeExtension get sized => _ContextSizeExtension(this);
-
-  @Deprecated('Use sized.emptySizedWidthBoxLow instead')
-  Widget get emptySizedWidthBoxLow => const SpaceSizedWidthBox(width: 0.01);
-
-  @Deprecated('Use sized.emptySizedWidthBoxLow3x instead')
-  Widget get emptySizedWidthBoxLow3x => const SpaceSizedWidthBox(width: 0.03);
-  @Deprecated('Use sized.emptySizedWidthBoxNormal instead')
-  Widget get emptySizedWidthBoxNormal => const SpaceSizedWidthBox(width: 0.05);
-  @Deprecated('Use sized.emptySizedWidthBoxHigh instead')
-  Widget get emptySizedWidthBoxHigh => const SpaceSizedWidthBox(width: 0.1);
-  @Deprecated('Use sized.emptySizedHeightBoxLow instead')
-  Widget get emptySizedHeightBoxLow => const SpaceSizedHeightBox(height: 0.01);
-  @Deprecated('Use sized.emptySizedHeightBoxLow3x instead')
-  Widget get emptySizedHeightBoxLow3x =>
-      const SpaceSizedHeightBox(height: 0.03);
-  @Deprecated('Use sized.emptySizedHeightBoxNormal instead')
-  Widget get emptySizedHeightBoxNormal =>
-      const SpaceSizedHeightBox(height: 0.05);
-  @Deprecated('Use sized.emptySizedHeightBoxHigh instead')
-  Widget get emptySizedHeightBoxHigh => const SpaceSizedHeightBox(height: 0.1);
-
-  @Deprecated('Use sized.height instead')
-  double get height => mediaQuery.size.height;
-  @Deprecated('Use sized.width instead')
-  double get width => mediaQuery.size.width;
-
-  @Deprecated(
-    'Use size.lowValue instead. ',
-  )
-  double get lowValue => height * 0.01;
-  @Deprecated('Use sized.normalValue instead')
-  double get normalValue => height * 0.02;
-  @Deprecated('Use sized.mediumValue instead')
-  double get mediumValue => height * 0.04;
-  @Deprecated('Use sized.highValue instead')
-  double get highValue => height * 0.1;
-
-  @Deprecated('Use sized.dynamicWidth instead')
-  double dynamicWidth(double val) => width * val;
-  @Deprecated('Use sized.dynamicHeight instead')
   double dynamicHeight(double val) => height * val;
 }