+++ lib/src/utility/index.dart
@@ -1,3 +1,4 @@
 export 'bundle/index.dart';
+export 'logger/custom_logger.dart';
 export 'device_utility.dart';
-export 'page_animation/index.dart';
+export 'link_preview/index.dart';