+++ lib/src/utility/maps_utility.dart
@@ -1,6 +1,4 @@
-import 'package:flutter/foundation.dart';
 import 'package:kartal/kartal.dart';
-import 'package:kartal/src/utility/core/custom_logger.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 
 typedef LaunchUrlCallBack = Future<bool> Function(
@@ -10,11 +8,14 @@ typedef LaunchUrlCallBack = Future<bool> Function(
   String? webOnlyWindowName,
 });
 
-mixin MapsUtility {
+final class MapsUtility {
+  const MapsUtility._();
+
   /// The function opens Apple Maps with a specified query.
   ///
   /// Args:
-  ///   query (String): A string representing the location or address you want to search for in Apple Maps.
+  ///   query (String): A string representing the location or
+  ///   address you want to search for in Apple Maps.
   static Future<bool> openAppleMapsWithQuery(
     String query, {
     LaunchUrlCallBack? callBack,
@@ -26,7 +27,7 @@ mixin MapsUtility {
       return callBack?.call(appleMapsWithQuery) ??
           launchUrlString(appleMapsWithQuery);
     } catch (error) {
-      CustomLogger.error<MapsUtility>(error);
+      CustomLogger.showError<MapsUtility>(error);
 
       return false;
     }
@@ -35,7 +36,8 @@ mixin MapsUtility {
   /// The function opens Google Maps with a specified query.
   ///
   /// Args:
-  ///   query (String): A string representing the search query to be used in Google Maps.
+  ///   query (String): A string representing the search query
+  /// to be used in Google Maps.
   static Future<bool> openGoogleMapsWithQuery(
     String query, {
     LaunchUrlCallBack? callBack,
@@ -47,7 +49,7 @@ mixin MapsUtility {
       return callBack?.call(googleMapsWithQuery) ??
           launchUrlString(googleMapsWithQuery);
     } catch (error) {
-      CustomLogger.error<MapsUtility>(error);
+      CustomLogger.showError<MapsUtility>(error);
       return false;
     }
   }
@@ -55,7 +57,8 @@ mixin MapsUtility {
   /// The function opens Google Web Maps with a specified query.
   ///
   /// Args:
-  ///   query (String): A string representing the search query to be used in Google Maps.
+  ///   query (String): A string representing the search query to be used in
+  /// Google Maps.
   static Future<bool> openGoogleWebMapsWithQuery(String query) async {
     final googleMapsWithQuery =
         '${KartalAppConstants.GOOGLE_MAPS_WEB_LINK}$query';
@@ -63,7 +66,7 @@ mixin MapsUtility {
     try {
       return launchUrlString(googleMapsWithQuery);
     } catch (error) {
-      CustomLogger.error<MapsUtility>(error);
+      CustomLogger.showError<MapsUtility>(error);
       return false;
     }
   }