+++ test/utility/bundle_decode_test.dart
@@ -1,10 +1,8 @@
 // ignore_for_file: avoid_relative_lib_imports
 
-import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:kartal/src/utility/bundle/bundle_decoder.dart';
 
-import '../../example/lib/samples/model/false_model.dart';
 import '../../example/lib/samples/model/post.dart';
 import '../mock/mock_bundle_asset.dart';